refactor(e2e): Added GH reporter base class
